requirements.txt
certifi==2023.5.7
setuptools==60.2.0
packaging==23.1
pip==23.1.2
websockets==11.0.3
click==8.1.3
pytest==7.3.1
PyYAML==6.0
typing-extensions==4.5.0
asgiref==3.7.0
anyio==3.6.2
idna==3.4
MarkupSafe==2.1.2
SQLAlchemy==2.0.15
greenlet==2.0.2
iniconfig==2.0.0
alembic==1.11.1
pydantic==1.10.7
pydantic==1.10.7
Mako==1.2.4
pluggy==1.0.0
python-dotenv==1.0.0
python-dotenv==1.0.0
wheel==0.37.1
asyncpg==0.27.0
starlette==0.27.0
httpx==0.24.1
uvicorn==0.22.0
pytest-asyncio==0.21.0
watchfiles==0.19.0
httpcore==0.17.2
h11==0.14.0
httptools==0.5.0
colorama==0.4.6